The homepage view for the phone display was created and animations were added.
