+++ 2. Application/src/components/WelcomeComp/WelcomeComp.css
@@ -1,7 +1,12 @@
+.rotate-pie {
+  transform-origin: 50% 0;
+  transform-box: fill-box;
+}
+
 .main-draw {
   display: block;
   width: 100%;
-  height: 13em;
+  height: 14em;
   fill: none;
   stroke: #c7e9ff;
   text-align: center;
